utils/build_stats_functions/build_profile/requirements.txt
functions-framework==3.8
google-cloud-bigquery==2.34.4
google-cloud-logging==3.3.0
google-cloud-storage==2.5.0 
